game process now run on a separate thread
