+++ babylonia-terminal-gui/src/manager.rs
@@ -3,8 +3,13 @@ use babylonia_terminal_sdk::{
     game_manager::GameManager,
 };
 use log::error;
-use relm4::tokio::sync::OnceCell;
+use relm4::{
+    tokio::{self, sync::OnceCell},
+    Worker,
+};
 use wincompatlib::prelude::Proton;
+
+use crate::ui;
 
 static PROTON: OnceCell<Proton> = OnceCell::const_new();
 
@@ -30,4 +35,39 @@ pub async fn run_game() {
     }
 
     GameManager::start_game(&proton, game_dir.unwrap(), None, false).await;
+}
+
+#[derive(Debug)]
+pub enum HandleGameProcessMsg {
+    RunGame,
+}
+
+pub struct HandleGameProcess;
+
+impl Worker for HandleGameProcess {
+    type Init = ();
+
+    type Input = HandleGameProcessMsg;
+
+    type Output = ui::MainWindowMsg;
+
+    fn init(_init: Self::Init, _sender: relm4::ComponentSender<Self>) -> Self {
+        Self
+    }
+
+    fn update(&mut self, message: Self::Input, sender: relm4::ComponentSender<Self>) {
+        match message {
+            HandleGameProcessMsg::RunGame => {
+                tokio::runtime::Builder::new_current_thread()
+                    .enable_all()
+                    .build()
+                    .unwrap()
+                    .block_on(async {
+                        sender.output(ui::MainWindowMsg::SetIsGameRunning(true));
+                        run_game().await;
+                        sender.output(ui::MainWindowMsg::SetIsGameRunning(false));
+                    });
+            }
+        }
+    }
 }
